% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkeletonCode.R
\name{MoveOneStep}
\alias{MoveOneStep}
\title{Generate movement from one grid to the next}
\usage{
MoveOneStep(starting.prob, grid, rate)
}
\arguments{
\item{starting.prob}{Same size as grid, but has probability of that individual
being in each cell; at tips, presumably 1 in one cell, 0 in all others, but
as you go down this spreads out}

\item{grid}{Two dimentional matrix of available spaces at this time step}

\item{rate}{Probability of trying a move to a neighboring cell. Right now, rate same
for all individuals, but could make trait dependent in future}
}
\value{
A matrix with probability of encountering the species in each grid cell
}
\description{
Generate movement from one grid to the next
}

